+++ packages/ndk/lib/data_layer/repositories/nostr_transport/websocket_client_nostr_transport_factory.dart
@@ -1,5 +1,5 @@
 import 'package:ndk/data_layer/repositories/nostr_transport/websocket_client_nostr_transport.dart';
-import 'package:ndk_web_socket_client/ndk_web_socket_client.dart';
+import '../../../src/web_socket_client/web_socket_client.dart';
 
 import '../../../domain_layer/repositories/nostr_transport.dart';
 import '../../../shared/helpers/relay_helper.dart';