+++ packages/ndk/test/data_layer/web_socket_client/src/web_socket_test.dart
@@ -5,7 +5,7 @@ import 'package:mocktail/mocktail.dart';
 import 'package:test/test.dart';
 import 'package:web_socket_channel/io.dart';
 import 'package:web_socket_channel/web_socket_channel.dart';
-import 'package:ndk_web_socket_client/ndk_web_socket_client.dart';
+import 'package:ndk/src/web_socket_client/web_socket_client.dart';
 
 class _MockBackoff extends Mock implements Backoff {}