+++ packages/web_socket_client/lib/ndk_web_socket_client.dart
@@ -1,5 +1,0 @@
-/// A reconnecting WebSocket client with native compression control.
-library;
-
-export 'package:web_socket_client/web_socket_client.dart' hide WebSocket;
-export 'src/web_socket.dart' show WebSocket;